// OmkarFCUBEAPP/src/app/models/ewaybillextmodel.ts
export class Ewaybillextmodel {
    bookedAt: string = "";
    bookingDate: string = "";
    gcNoteNo: string = "";
    ewayBillNo: string = "";
    ewayBillDate: string = "";
    ewayBillExpDate: string = "";
    vehicleNo: string = "";
    fromLocation: string = "";
    fromPin : string = "";
    destination: string = "";
    partyName: string = "";
    consignor: string = "";
    consignee: string = "";
    accountCity : string = "";
    cnorState : string = "";
    accountAddress1: string = "";
    accountAddress2: string = "";
    accountAddress3: string = ""; 
    kMS: string = ""; 
    state: string = "";
    mode:string = ""; 
    reason:string = ""; 
    remarks:string = ""; 
    consignmentStatus:string = ""; 
    loggedInUser:string = ""; 
}